+++ src/hooks/actor.ts
@@ -7,7 +7,11 @@ import {handleBaseRule, handleComplexRule} from "./message";
 export class RulePreUpdateActorHook implements PreUpdateActorHook {
     listen(actor: Actor, data: object, o: object) {
         const options = actor.getRollOptions()
-        const mm = new MessageForHandling(undefined, undefined, new Set(options), o)
+        const mm = new MessageForHandling(undefined, undefined, new Set(options), {
+            origin: {
+                actor: actor?.uuid,
+            }
+        })
         mm.mainActor = actor;
         mm.data = foundry.utils.deepClone(data);